+++ ratpack-dropwizard-metrics/src/test/groovy/ratpack/dropwizard/metrics/MetricsSpec.groovy
@@ -32,10 +32,20 @@ import org.junit.rules.TemporaryFolder
 import org.slf4j.Logger
 import ratpack.dropwizard.metrics.internal.PooledByteBufAllocatorMetricSet
 import ratpack.dropwizard.metrics.internal.UnpooledByteBufAllocatorMetricSet
+import ratpack.error.ServerErrorHandler
+import ratpack.error.internal.DefaultDevelopmentErrorHandler
 import ratpack.exec.Blocking
+import ratpack.exec.ExecController
 import ratpack.exec.Promise
+import ratpack.groovy.handling.GroovyChain
+import ratpack.groovy.test.embed.GroovyEmbeddedApp
+import ratpack.http.client.HttpClient
+import ratpack.http.client.internal.DefaultHttpClient
+import ratpack.test.embed.EmbeddedApp
 import ratpack.test.internal.RatpackGroovyDslSpec
 import ratpack.websocket.RecordingWebSocketClient
+import spock.lang.AutoCleanup
+import spock.util.concurrent.BlockingVariable
 import spock.util.concurrent.PollingConditions
 
 import java.time.Duration
@@ -49,6 +59,20 @@ class MetricsSpec extends RatpackGroovyDslSpec {
 
   @Rule
   TemporaryFolder reportDirectory
+
+  @AutoCleanup
+  EmbeddedApp otherApp
+
+  EmbeddedApp otherApp(@DelegatesTo(value = GroovyChain, strategy = Closure.DELEGATE_FIRST) Closure<?> closure) {
+    otherApp = GroovyEmbeddedApp.of {
+      registryOf { add ServerErrorHandler, new DefaultDevelopmentErrorHandler() }
+      handlers(closure)
+    }
+  }
+
+  URI otherAppUrl(String path = "") {
+    new URI("$otherApp.address$path")
+  }
 
   def setup() {
     SharedMetricRegistries.clear()
@@ -762,4 +786,128 @@ class MetricsSpec extends RatpackGroovyDslSpec {
     0 * reporter.onTimerAdded("foo.get-requests", !null)
   }
 
+  def "can use prometheus metrics endpoint"() {
+    given:
+    bindings {
+      module new DropwizardMetricsModule(), { it.prometheusCollection(true) }
+    }
+    handlers { MetricRegistry metrics ->
+
+      metrics.register("fooGauge", new com.codahale.metrics.Gauge<Integer>() {
+        @Override
+        Integer getValue() {
+          2
+        }
+      })
+
+      get {
+        metrics.meter("fooMeter").mark()
+        metrics.counter("fooCounter").inc()
+        metrics.histogram("fooHistogram").update(metrics.counter("fooCounter").count)
+        render "foo"
+      }
+
+      get("admin/metrics-report", new MetricsPrometheusHandler())
+    }
+
+    when:
+    get("/")
+    def metrics = getText("admin/metrics-report").split("\n")
+
+    then:
+    metrics.contains("fooGauge 2.0")
+    metrics.contains("_2xx_responses 1.0")
+    metrics.contains("fooCounter 1.0")
+    metrics.contains("fooHistogram{quantile=\"0.5\",} 1.0")
+    metrics.contains("fooHistogram{quantile=\"0.75\",} 1.0")
+    metrics.contains("fooHistogram{quantile=\"0.95\",} 1.0")
+    metrics.contains("fooHistogram{quantile=\"0.98\",} 1.0")
+    metrics.contains("fooHistogram{quantile=\"0.99\",} 1.0")
+    metrics.contains("fooHistogram{quantile=\"0.999\",} 1.0")
+    metrics.contains("fooHistogram_count 1.0")
+    metrics.find { it.startsWith("root_get_requests{quantile=\"0.5\",}") }
+    metrics.find { it.startsWith("root_get_requests{quantile=\"0.75\",}") }
+    metrics.find { it.startsWith("root_get_requests{quantile=\"0.95\",}") }
+    metrics.find { it.startsWith("root_get_requests{quantile=\"0.98\",}") }
+    metrics.find { it.startsWith("root_get_requests{quantile=\"0.99\",}") }
+    metrics.find { it.startsWith("root_get_requests{quantile=\"0.999\",}") }
+    metrics.contains("root_get_requests_count 1.0")
+    metrics.contains("fooMeter_total 1.0")
+
+  }
+
+  def "it should report http client metrics"() {
+    given:
+    System.setProperty("ratpack.metrics.httpclient.enabled", "true")
+    System.setProperty("ratpack.metrics.httpclient.pollingFrequency", "1")
+
+    MetricRegistry registry
+    String ok = 'ok'
+    def result = new BlockingVariable<String>()
+    def httpClient = HttpClient.of { DefaultHttpClient.Spec spec ->
+      spec.poolSize(0)
+      spec.enableMetricsCollection(true)
+    }
+
+    bindings {
+      bindInstance(HttpClient, httpClient)
+      module new DropwizardMetricsModule()
+    }
+
+    otherApp {
+      get {
+        Blocking.get({
+          return result.get()
+        })
+          .onError(it.&error)
+          .then(it.&render)
+      }
+    }
+
+    handlers { MetricRegistry metrics ->
+      registry = metrics
+      get {
+        ExecController execController = it.get(ExecController)
+        execController.fork().start({
+          httpClient.get(otherAppUrl())
+            .then({ val ->
+            assert val.body.text == ok
+          })
+        })
+        render ok
+      }
+    }
+
+    when:
+    text == ok
+
+    then:
+    polling.within(2) {
+      assert registry.getGauges().get('httpclient.total.active.connections').value == 1
+      assert registry.getGauges().get('httpclient.total.idle.connections').value == 0
+      assert registry.getGauges().get('httpclient.total.connections').value == 1
+      assert registry.getGauges().get("httpclient.${otherAppUrl().host}.total.active.connections").value == 1
+      assert registry.getGauges().get("httpclient.${otherAppUrl().host}.total.idle.connections").value == 0
+      assert registry.getGauges().get("httpclient.${otherAppUrl().host}.total.connections").value == 1
+
+
+    }
+
+    when:
+    result.set(ok)
+
+    then:
+    polling.within(2) {
+      assert registry.getGauges().get('httpclient.total.active.connections').value == 0
+      assert registry.getGauges().get('httpclient.total.idle.connections').value == 0
+      assert registry.getGauges().get('httpclient.total.connections').value == 0
+      assert registry.getGauges().get("httpclient.${otherAppUrl().host}.total.active.connections").value == 0
+      assert registry.getGauges().get("httpclient.${otherAppUrl().host}.total.idle.connections").value == 0
+      assert registry.getGauges().get("httpclient.${otherAppUrl().host}.total.connections").value == 0
+    }
+
+    cleanup:
+    System.setProperty("ratpack.metrics.httpclient.enabled", "false")
+    System.setProperty("ratpack.metrics.httpclient.pollingFrequency", "30")
+  }
 }